Create Palindrome number
This code works by first checking if x is negative. If it is, then it is not a palindrome. And now checks if x is a single digit, if it is a single then returns true. Otherwise, the code reverses x by repeatedly multiplying reversed x by 10 and adding the last digit of x. Once x has been reversed, the code compares it to x to see if they are equal. If they are, then x is a palindrome.