Create Roman to integer
This code is designed to convert a Roman numeral string into its corresponding integer value. It achieves this by iterating through the string in reverse order, using a for loop. Starting from the last character of the Roman numeral, the code checks if the current value is greater than or equal to the previous value encountered. If it is, the current value is added to the result. Otherwise, it is subtracted from the result. The previous value is then updated to the current value for the next iteration. This process continues until the loop reaches the first character of the Roman numeral. Ultimately, the function returns the resulting integer value.