Create Longest common prefix
The code first checks if the input array is empty. If it is, it returns an empty string. Otherwise, it finds the minimum length of all the strings in the input array. It then initializes the prefix to the first character of the first string in the input array. It then iterates over the first min_len characters of each string in the input array. For each character, it checks if all the strings in the input array have the same character at the current index. If so, it continues to the next character. Otherwise, it breaks out of the loop. Finally, it returns the prefix.